fix small bug in peek. use Float64Array for memory by default for improvement timing/phase accuracy.

+++ js/gen.js
@@ -83,7 +83,7 @@ let gen = {
    * ... the generated function will have a signature of ( abs, p0 ).
    */
   
-  createCallback( ugen, mem, debug = false, shouldInlineMemory=false, memType = Float32Array ) {
+  createCallback( ugen, mem, debug = false, shouldInlineMemory=false, memType = Float64Array ) {
     let isStereo = Array.isArray( ugen ) && ugen.length > 1,
         callback, 
         channel1, channel2
